// src/app/distrital/interfaces/computo.interface.ts
export interface Computo {
  spen_1:           boolean;
  spen_2:           boolean;
  spen_3:           boolean;
  spen_4:           boolean;
  spen_5:           boolean;
  // quorum_asistencia:string;
  mc_prensa:        boolean;
  mc_radio:         boolean;
  mc_tv:            boolean;
  fecha_sesion:     string;
  hora_sesion:      string;
  observaciones:    string;
}